fixed adjmat makegraph

+++ Constants.h
@@ -1,6 +1,6 @@
 #include <iostream>
 #pragma once
-
+#include <cfloat>
 #define PROTECTED
 #define PUBLIC
 #define PRIVATE
@@ -15,11 +15,12 @@ namespace Graph {
         int i_start;
         int j_end;
         double weight;
+
     } arc;
 
     typedef struct ITEM {
         int data;
-        int key;
+        double key;
     } item;
